+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.7.1 (Unreleased)
 - **[Feature]** Support context-base configuration with low-level topic settings alterations producer variants.
+- [Enhancement] Prefix random default `SecureRandom.hex(6)` producers ids with `waterdrop-hex` to indicate type of object.
 
 ## 2.7.0 (2024-04-26)